feat: Add experimental ESM externals configuration to next.config
This commit adds the experimental configuration option `esmExternals` to the `next.config` file. The value is set to 'loose'. This configuration is intended to improve the performance of the page load by lazy loading images.

Refactor: Remove unused page component for channel points

This commit removes the unused page component `LoginPage` from the `src/app/login` directory. The component was no longer being used and can be safely deleted.

Fix: Update banned songs layout and page components

This commit updates the layout and page components related to the banned songs feature. It fixes some issues and improves the overall user experience.

chore: Update npm dependencies and add @uidotdev/usehooks package

This commit updates the npm dependencies to the latest stable versions and adds the `@uidotdev/usehooks` package to the project.

chore: Update Twitch login scopes in login actions

This commit updates the Twitch login scopes in the `login` action file. It ensures that the necessary scopes are requested when authenticating with Twitch.